Split components: JS
