+++ src/Card/Card.js
@@ -1,0 +1,18 @@
+import React from 'react';
+const Card = (props) => {
+    return (
+      <div
+        id = {props.cardProperties.id}
+        className = {`cards${props.numberOfCards} card` + (props.cardProperties.opened?" flip":" scale")}
+        onClick = {() => props.onClick(props.cardProperties.id)}
+        style = {{
+          order: props.cardProperties.positionOnField,
+          visibility: props.cardProperties.hidden ? "hidden" : "visible",
+        }}
+      >
+        <img src=  "img/js-badge.svg" alt = {props.cardProperties.id} className = "backImg"/>
+        <img src={`img/${props.cardProperties.hiddenValue}.svg`} alt = {props.cardProperties.id} className = "frontImg"/>
+      </div>
+    );
+  }
+  export default Card;
