+++ src/CardsContainer/CardsContainer.js
@@ -1,0 +1,88 @@
+import React, { useState } from 'react';
+import Timer from './Timer/Timer.js';
+import Card from '../Card/Card.js';
+
+const CardsContainer = (props) => {
+    const cardsPropertiesStart = Array(props.numberOfCards).fill(0).map((_, index) => ({
+      id:index,
+      hiddenValue:Math.floor(index/2),
+      hidden:false,
+      opened:false,
+      positionOnField:Math.floor(Math.random()*50)
+    }));
+  
+    let cardsPropertiesNew = [];
+  
+    const [cardsProperties, setCardsProperties] = useState(cardsPropertiesStart);
+  
+    const [fieldIsLocked, setFieldIsLocked] = useState(false);
+  
+    const getVisibleCardsNum = () => cardsPropertiesNew.filter(card => !(card.hidden)).length;
+  
+    const hideCards = () => {
+      setTimeout(() => {
+        const cardsForHide = getOpenedCards();
+        cardsForHide.forEach(card => card.hidden = true);
+        cardsForHide.forEach(card => card.opened = false);
+        setCardsProperties(cardsPropertiesNew);
+        if (!getVisibleCardsNum()) props.setTimerIsActive(false);
+        setFieldIsLocked(false);
+      }, 1000);
+    }
+  
+    const isCardOpened = (id) => cardsPropertiesNew[id].opened;
+  
+    const cardOnClick = (id) => {
+      if (fieldIsLocked) return; 
+      cardsPropertiesNew = [...cardsProperties];
+      switch (getOpenedCards().length) {
+        case 0:
+          cardsPropertiesNew[id].opened = true;
+          setCardsProperties(cardsPropertiesNew);
+          return;
+        case 1:
+          if (isCardOpened(id)) return; 
+          cardsPropertiesNew[id].opened = true;
+          setCardsProperties(cardsPropertiesNew);
+          setFieldIsLocked(true);
+          if(areOpenedCardsEqual()) {
+            hideCards();
+            return;
+          };
+          setTimeout(() => {
+            getOpenedCards().forEach(card => card.opened = false);
+            setCardsProperties(cardsPropertiesNew);
+            setFieldIsLocked(false);
+          }, 600);
+          return;
+        default:
+          return;
+      }
+    }
+  
+    const getOpenedCards = () => (cardsPropertiesNew.filter(card => card.opened));
+  
+    const areOpenedCardsEqual = () => {
+      const openedCards = getOpenedCards();
+      return openedCards[0].hiddenValue === openedCards[1].hiddenValue;
+    };
+  
+    return (
+      <main id = "cardsContainer">
+        <div id = "innerContainer">
+        <Timer 
+          timerValue = {props.timerValue}
+        />
+          {cardsProperties.map((cardProperties, index) =>(
+            <Card
+              key = {index}
+              cardProperties = {cardProperties}
+              onClick = {(id) => cardOnClick(id)}
+              numberOfCards = {props.numberOfCards}
+            />
+          ))}
+        </div>
+      </main>
+    );
+  }
+  export default CardsContainer;
