+++ src/ChooseDifficulty/ChooseDifficulty.js
@@ -1,0 +1,28 @@
+import React from 'react';
+const ChooseDifficulty = (props) => {
+    return (
+      <div className = "modal" style = {{visibility: props.isDifficultyVisible ? "visible" : "hidden",}}>
+        <div className = "modalContent">
+          <p>Please, choose a level:</p>
+          <div className = "radioContainer">
+              <input type = "radio" name = "difficulty" id = "Easy" value = "12" onChange = {props.difficultyOnChange}>
+              </input>
+              <label htmlFor = "Easy">Easy</label>
+          </div>
+          <div className = "radioContainer">
+              <input type = "radio" name = "difficulty" id = "Medium" value = "20" onChange = {props.difficultyOnChange}>
+              </input>
+              <label htmlFor="Medium">Medium</label>
+          </div>
+          <div className="radioContainer">
+              <input type="radio" name="difficulty" id="Hard" value="30" onChange = {props.difficultyOnChange}>
+              </input>
+              <label htmlFor="Hard">Hard</label>
+          </div>
+            <input type="submit" value="Start" id="submit" disabled = {!props.isButtonEnabled} onClick = {props.buttonOnClick}>
+            </input>
+        </div>
+      </div>
+    )
+  }
+  export default ChooseDifficulty;
