+++ src/Timer/Timer.js
@@ -1,0 +1,7 @@
+import React from 'react';
+const Timer = (props) => {
+    return (
+      <div id="timer">{props.timerValue}</div>
+    )
+  }
+export default Timer;
